requirements.txt
absl-py==0.11.0
appdirs==1.4.4
argon2-cffi==20.1.0
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.1
bs4==0.0.1
cachetools==4.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cloudpickle==1.6.0
cssselect==1.1.0
cycler==0.10.0
debugpy==1.3.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
fake-useragent==0.1.11
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gym==0.18.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.1.1
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==5.1.3
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parse==1.19.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyee==8.1.0
pyglet==1.5.0
Pygments==2.9.0
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2020.5
pyzmq==20.0.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.7.1
six==1.15.0
soupsieve==2.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
#torch==0.4.1
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.5.0